pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
    "hypercorn>=0.17.3",
    "gunicorn>=23.0.0",
    "flask>=3.1.0",
    "requests>=2.32.3",
    "numpy>=2.2.5",
    "neo4j>=5.28.1",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "twilio>=9.6.1",
    "pillow>=11.2.1",
    "google-auth-httplib2>=0.2.0",
    "werkzeug>=3.1.3",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
]
